UserInfoSchema が未初期化って怒られるのでファイル分け 'UserInfoSchema' before initialization
