+++ frontend/src/features/user/types/UserDto.ts
@@ -1,8 +1,5 @@
 import { z } from 'zod';
 
-import { chatChannelSchema } from '@/features/chat/types/chatChannelDto';
-import { ReceiveInviteChatRoomDtoSchema } from '@/features/chat/types/InviteChatRoomDto';
-
 export const SignUpDtoSchema = z.object({
   email: z.string().email(),
   username: z.string(),
@@ -24,13 +21,3 @@ export type UserInfo = z.infer<typeof UserInfoSchema>;
 
 export const UserInfoArrSchema = UserInfoSchema.array();
 export type UserInfoArr = z.infer<typeof UserInfoArrSchema>;
-
-export const CurrentUserSchema = z.object({
-  userInfo: UserInfoSchema,
-  friends: UserInfoSchema.array(),
-  friendRequests: UserInfoSchema.array(),
-  blockUsers: UserInfoSchema.array(),
-  joinedRooms: chatChannelSchema.array(),
-  receiveInviteChatRooms: ReceiveInviteChatRoomDtoSchema.array(),
-});
-export type CurrentUser = z.infer<typeof CurrentUserSchema>;
