+++ frontend/src/stores/jotai.ts
@@ -1,7 +1,7 @@
 import { atom } from 'jotai';
 import { Socket } from 'socket.io-client';
 
-import { CurrentUser } from '../features/user/types/UserDto';
+import { CurrentUser } from '@/features/user/types/CurrentUser';
 
 // TODO : ファイル名をかもうちょい考える, いったん移動
 export const currentUserAtom = atom<CurrentUser | undefined>(undefined);
